modify service for admin binhpnt commit 18112022

+++ API/Controllers/ServiceController.cs
@@ -26,12 +26,19 @@ namespace API.Controllers
         }
 
         [HttpGet("get/service")]
-        public async Task<IEnumerable<ServiceDto>> GetService(Guid? id, Guid airlineId)
+        public async Task<Response<IEnumerable<ServiceDto>>> GetService(Guid? id, Guid? airlineId)
         {
             var dp_params = new DynamicParameters();
+            var response = new Response<IEnumerable<ServiceDto>>();
             dp_params.Add("@Id", id, DbType.Guid); 
             dp_params.Add("@AirlineId", airlineId, DbType.Guid);
-            return await _db.GetAll<ServiceDto>("GetService", dp_params);
+            var newData = await _db.GetAll<ServiceDto>("GetService", dp_params);
+            if(newData != null)
+            {
+                response.Code = 200;
+                response.Data = newData;
+            }
+            return response;
         }
 
         [HttpPost("create/service")]
@@ -47,15 +54,24 @@ namespace API.Controllers
             {
                 string sqlCommand = "insert into Service(ServiceName, Cost, AirlineId)" +
                     "values (@ServiceName, @Cost, @AirlineId)";
-                if(await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text) == 1)
+                try
                 {
-                    sqlCommand = "select top 1 * from Service order by Id desc";
-                    var newData = await db.QueryFirstAsync<ServiceDto>(sqlCommand, null, null, null, CommandType.Text);
-                    if(newData != null)
+                    if(await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text) == 1)
                     {
-                        response.Code = 200;
-                        response.Data = newData;
+                        sqlCommand = "select s.Id, s.ServiceName, s.Cost, s.AirlineId, a.AirlineName from Airline a " +
+                            "inner join Service s on a.Id = s.AirlineId order by Id desc";
+                        var newData = await db.QueryFirstAsync<ServiceDto>(sqlCommand, null, null, null, CommandType.Text);
+                        if(newData != null)
+                        {
+                            response.Code = 200;
+                            response.Data = newData;
+                        }
                     }
+                } catch(Exception ex)
+                {
+                    response.Code = 500;
+                    response.Data = null;
+                    return response;
                 }
             };
             return response;
@@ -89,16 +105,37 @@ namespace API.Controllers
         }
 
         [HttpDelete("delete/service")]
-        public async Task<ActionResult> DeleteService(Guid id)
+        public async Task<Response<ServiceDto>> DeleteService(Guid id)
         {
             var dp_params = new DynamicParameters();
+            Response<ServiceDto> response = new Response<ServiceDto>();
+
             dp_params.Add("@Id", id, DbType.Guid);
             using (IDbConnection db = new SqlConnection(_config.GetConnectionString("DefaultConnection")))
             {
-                string sqlCommand = "delete from Service where Id = @Id";
-                await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text);
+                string sqlCommand = "select s.Id, s.ServiceName, s.Cost, s.AirlineId, a.AirlineName from Airline a " +
+                            "inner join Service s on a.Id = s.AirlineId order by Id desc";
+                var oldData = await db.QueryFirstAsync<ServiceDto>(sqlCommand, dp_params, null, null, CommandType.Text);
+
+                if (oldData != null)
+                {
+                    try
+                    {
+                        sqlCommand = "delete from Service where Id = @Id";
+                        await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text);
+
+                        response.Code = 200;
+                        response.Data = oldData;
+                    }
+                    catch (Exception ex)
+                    {
+                        response.Code = 500;
+                        response.Data = null;
+                        return response;
+                    }
+                }
             };
-            return Ok("Success");
+            return response;
         }
     }
 }
